% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starchart-package.R
\docType{package}
\name{starchart-package}
\alias{starchart}
\alias{starchart-package}
\title{starchart: 'R-Universe' 'API'}
\description{
Client to access the 'R-Universe' 'API'.
}
\author{
\strong{Maintainer}: Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

}
\keyword{internal}
